# 管理后台/packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AsyncAccepted": {
        "properties": {
          "job_id": {
            "type": "string"
          },
          "resource_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "description": "**不许返回「已完成」来冒充同步执行**",
            "enum": [
              "排队中",
              "执行中",
              "已完成",
              "失败",
              "取消请求中",
              "已取消",
              "状态待核实"
            ],
            "type": "string"
          },
          "status_url": {
            "type": "string"
          },
          "trace_id": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "job_id",
          "status",
          "resource_id",
          "status_url",
          "trace_id"
        ],
        "type": "object"
      },
      "Error": {
        "properties": {
          "advice": {
            "description": "下一步该干什么。写不出来说明还没搞清失败原因",
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "field_errors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "trace_id": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "advice"
        ],
        "type": "object"
      },
      "ListEnvelope": {
        "properties": {
          "items": {
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "total": {
            "description": "未知时必须是 null。**不许用 0 表示「我没数」**",
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "next_cursor",
          "total"
        ],
        "type": "object"
      }
    },
...
  },
  "info": {
    "description": "从 `services/api/app/contract/` 生成,**不要手改**。\n\n契约在前、代码在后:让 FastAPI 从 handler 反推 OpenAPI 的话,**契约就变成了实现的影子** —— handler 漏了一个鉴权,OpenAPI 跟着漏,而它看起来仍然是一份完整的契约。\n\n⚠️ 这是**待实施契约**,不代表已经有可用的服务。",
    "title": "澜绣云裳 AI 管理后台 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/knowledge-bases": {
      "get": {
        "operationId": "get_knowledge-bases",
        "parameters": [
          {
            "description": "**服务端重新校验项目权限与对象范围** —— 改这个 ID 绕不过授权(§19.1)",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEnvelope"
                }
              }
            },
            "description": "分页列表"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "未认证"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "无权"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "不存在,或按策略隐藏存在性"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "冲突(版本被人改过 / 幂等键同键不同请求)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "输入无效"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "限额"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "服务问题"
          }
        },
        "summary": "知识库列表",
        "tags": [
          "知识与 RAG"
        ],
        "x-字段权限": [],
        "x-审计": false,
        "x-权限": "查看有权配置"
      },
      "post": {
        "operationId": "post_knowledge-bases",
        "parameters": [
          {
            "description": "**服务端重新校验项目权限与对象范围** —— 改这个 ID 绕不过授权(§19.1)",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "已创建"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "未认证"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "无权"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "不存在,或按策略隐藏存在性"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "冲突(版本被人改过 / 幂等键同键不同请求)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "输入无效"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "限额"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "服务问题"
          }
        },
        "summary": "建知识库",
        "tags": [
          "知识与 RAG"
        ],
        "x-字段权限": [],
        "x-审计": true,
        "x-权限": "改 Prompt/知识候选"
      }
    },
...
    "/api/v1/projects/{project_id}/knowledge-bases/{id}/index-builds": {
      "get": {
        "description": "**成员数要和知识库的片段数对上** —— 对不上就是索引不完整,而一个不完整的索引检索时只是「少返回几条」,**不报错**",
        "operationId": "get_knowledge-bases_id_index-builds",
        "parameters": [
          {
            "description": "**服务端重新校验项目权限与对象范围** —— 改这个 ID 绕不过授权(§19.1)",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEnvelope"
                }
              }
            },
            "description": "分页列表"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "未认证"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "无权"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "不存在,或按策略隐藏存在性"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "冲突(版本被人改过 / 幂等键同键不同请求)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "输入无效"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "限额"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "服务问题"
          }
        },
        "summary": "某个知识库的索引构建",
        "tags": [
          "知识与 RAG"
        ],
        "x-字段权限": [],
        "x-审计": false,
        "x-权限": "查看有权配置"
      },
      "post": {
        "description": "**换了 Embedding 模型要先校验维度** —— 混进旧索引的向量算出来的相似度没有意义",
        "operationId": "post_knowledge-bases_id_index-builds",
        "parameters": [
          {
            "description": "**服务端重新校验项目权限与对象范围** —— 改这个 ID 绕不过授权(§19.1)",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "**必填**。相同键相同请求返回原任务;相同键不同请求返回 409。超时重发一次的代价是重复烧 GPU 或重复计费",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncAccepted"
                }
              }
            },
            "description": "已受理,去 status_url 看进度"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "未认证"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "无权"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "不存在,或按策略隐藏存在性"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "冲突(版本被人改过 / 幂等键同键不同请求)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "输入无效"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "限额"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "服务问题"
          }
        },
        "summary": "建索引",
        "tags": [
          "知识与 RAG"
        ],
        "x-字段权限": [],
        "x-审计": false,
        "x-权限": "改 Prompt/知识候选"
      }
    },
...
    "/api/v1/projects/{project_id}/retrieval-tests": {
      "post": {
        "description": "必须回:原问、改写、候选、融合分、选片、证据、Trace —— **整条链路**(§9.5「必须看到整条链路」)",
        "operationId": "post_retrieval-tests",
        "parameters": [
          {
            "description": "**服务端重新校验项目权限与对象范围** —— 改这个 ID 绕不过授权(§19.1)",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "**必填**。相同键相同请求返回原任务;相同键不同请求返回 409。超时重发一次的代价是重复烧 GPU 或重复计费",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncAccepted"
                }
              }
            },
            "description": "已受理,去 status_url 看进度"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "未认证"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "无权"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "不存在,或按策略隐藏存在性"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "冲突(版本被人改过 / 幂等键同键不同请求)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "输入无效"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "限额"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "服务问题"
          }
        },
        "summary": "检索实验室跑一次",
        "tags": [
          "知识与 RAG"
        ],
        "x-字段权限": [],
        "x-审计": false,
        "x-权限": "运行评测"
      }
    },
...
    "/api/v1/projects/{project_id}/uploads": {
      "post": {
        "description": "**服务端校验之后才算文件引用** —— 上传成功不等于内容可用",
        "operationId": "post_uploads",
        "parameters": [
          {
            "description": "**服务端重新校验项目权限与对象范围** —— 改这个 ID 绕不过授权(§19.1)",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "已创建"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "未认证"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "无权"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "不存在,或按策略隐藏存在性"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "冲突(版本被人改过 / 幂等键同键不同请求)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "输入无效"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "限额"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "服务问题"
          }
        },
        "summary": "要一个受限上传地址",
        "tags": [
          "知识与 RAG"
        ],
        "x-字段权限": [],
        "x-审计": false,
        "x-权限": "改 Prompt/知识候选"
      }
    },
...
  },
  "servers": [
    {
      "description": "同源",
      "url": "/"
    }
  ]
}